examples/config/requirements.txt
snowflake-connector-python[pandas]==3.15.0
python-dotenv==1.0.0
cryptography
pandas==2.2.1
pyarrow
mcp>=0.1.0
asyncio
pydantic>=2.0.0
sqlparse==0.4.4  # For SQL validation
cachetools==5.3.0  # For caching Cortex results
tenacity==8.2.3  # For retry logic
google-cloud-secret-manager==2.16.0  # For GCP secrets
fastapi==0.104.1  # For API server
uvicorn==0.24.0  # For ASGI server
